finalizando fluxo de gerenciamento de tenants

+++ app/Http/Controllers/TenantController.php
@@ -2,7 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Tenant;
+use App\Models\{
+    Tenant,
+    Role,
+    Permission,
+};
 use Illuminate\Http\Request;
 
 class TenantController extends Controller
@@ -10,15 +14,17 @@ class TenantController extends Controller
     public function getAllTenants()
     {
         try {
+            $this->authorize('to_manage_tenant');
             $data = Tenant::all();
             return response()->json($data, 200);
         } catch (\Throwable $th) {
-            return response()->json(['msg' => 'Falha ao tentar listar Tenants!'], 500);
+            return response()->json(/*['msg' => 'Falha ao tentar listar Tenants!']*/$th->getMessage(), 500);
         }
     }
 
     public function getTenant($id){
         try {
+            $this->authorize('to_manage_tenant');
             $data = Tenant::find($id);
             return response()->json($data, 200);
         } catch (\Throwable $th) {
@@ -29,6 +35,7 @@ class TenantController extends Controller
     public function createTenant(Request $request)
     {
         try {
+            $this->authorize('to_manage_tenant');
             Tenant::create($request->input());
             return response()->json(['msg' => 'Tenant cadastrado!'], 201);
         } catch (\Throwable $th) {
@@ -39,6 +46,7 @@ class TenantController extends Controller
     public function updateTenant(Request $request, $id)
     {
         try {
+            $this->authorize('to_manage_tenant');
             Tenant::where('id', $id)
                 ->update($request->all());
                 return response()->json(['msg' => 'Tenant atualizado!'], 200);
@@ -50,6 +58,7 @@ class TenantController extends Controller
     public function deleteTenant($id)
     {
         try {
+            $this->authorize('to_manage_tenant');
             Tenant::destroy($id);
             return response()->json(['msg' => 'Tenant removido!'], 200);
         } catch (\Throwable $th) {
